Added spectator mode to tournament games, small rework of GameFieldDrawer GameFieldDrawer is now it's own class, and has a setPosition and setSize function. The calFieldPos was reworked to perform well no matter the position and size (will probably not work well with very small sizes) Join room response was changed to sf::Uint16 to be able to receive challenge ID in it.

+++ src/gameplay/gameField.h
@@ -56,7 +56,7 @@ public:
 
     sf::Uint16 id;
     sf::Uint8 nextpiece, nprot, npcol;
-    short scale;
+    float scale;
     bool mouseover;
     sf::Uint8 datacount;
